Add BlendOver (like BlendNormal but with premultiplied alpha). Fix some compiler warnings. Fix a bug in the software blender when using INCLUDE_ALPHA_CHANNEL

+++ CrossFade.cpp
@@ -6,7 +6,7 @@ void CrossFade(dword *dst, dword stride,
                dword *srcA, dword *srcB,
                dword width, dword height, dword alpha)
 {
-  byte invAlpha  = 255 - alpha;
+  byte invAlpha  = (byte)(255 - alpha);
 
   for (dword y = 0; y < height; ++y)
   {
